+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ### Global Updates
 
-- [#IssueNumber](link-to-github-issue) 
+- [#1 - monthly budget summary: sheet to doc](https://github.com/official-carledwardfp/budget-tracker-appscript/issues/1)